added the backend part of login and the 401.html page

+++ app.py
@@ -1,7 +1,8 @@
 from flask import Flask, render_template, redirect, flash, request, url_for
 from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate 
-import flask_login
+from flask_migrate import Migrate
+import flask_login 
+from flask_login import LoginManager, UserMixin, login_required, current_user
 from datetime import datetime
 
 
@@ -12,6 +13,11 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
 login_manager = flask_login.LoginManager()
 login_manager.init_app(app)
+login_manager.login_view = 'login'
+
+@login_manager.user_loader
+def load_user(user_id):
+    return Users.query.get(int(user_id))
 
 
 db = SQLAlchemy(app)
@@ -19,7 +25,7 @@ migrate = Migrate(app, db)
 
 
 #----------------------TABLES-----------------------------
-class Users(db.Model):
+class Users(UserMixin, db.Model):
     id = db.Column(db.Integer, primary_key=True)
     Name = db.Column(db.String(128), nullable=True)
     Email = db.Column(db.String(128), unique=True, nullable=True)
@@ -95,8 +101,12 @@ def login():
             elif user_to_check.Password != password:
                 flash('Incorrect password, please try again...')
             else:
-                flash('Login successfull!')
-                return render_template('user_dashboard.html', user=user_to_check)
+                if user_to_check.Email == 'admin@mail.com':
+                    flash('Login successfull!')
+                    return redirect(url_for('admin'))
+                else:
+                    flash('Login successfull!')
+                    return render_template('user_dashboard.html', user=user_to_check)
         except Exception as e:
             flash(f'An error occured: {str(e)}')
             return redirect(url_for("index"))
@@ -104,19 +114,35 @@ def login():
 
 
 @app.route('/user/delete/<int:id>')
+@login_required
 def deleteuser(id):
     user = Users.query.get_or_404(id)
-    try:
-        db.session.delete(user)
-        db.session.commit()
-        flash('User has been deleted successfully. We are sad to see you leave')
-        return redirect(url_for('login'))
-    except Exception as e:
-        flash(f'An error has occured: {str(e)}')
-        return redirect(url_for("userlist"))      
+    if current_user.id == user.id:
+        try:
+            db.session.delete(user)
+            db.session.commit()
+            flash('User has been deleted successfully. We are sad to see you leave')
+            return redirect(url_for('login'))
+        except Exception as e:
+            flash(f'An error has occured: {str(e)}')
+            return redirect(url_for("userlist"))
+    else:
+        flash('You are unatuhorized to access this page')
+        return redirect(url_for('unauthorized'))      
+
+
+#----------------------------ADMIN--------------------------
+@app.route('/admin/dashboard')
+def admin():
+    return render_template('admin.html')
 
 
 #----------------------ERROR PAGES--------------------------
+@app.errorhandler(401)
+def unauthorized(e):
+    return render_template('401.html'), 401
+
+
 @app.errorhandler(404)
 def page_not_found(e):
     return render_template('404.html'), 404
